move file template
