+++ sky/utils/kubernetes_utils.py
@@ -9,11 +9,9 @@ from urllib.parse import urlparse
 import jinja2
 import yaml
 
-import sky
 from sky import exceptions
 from sky import sky_logging
 from sky.adaptors import kubernetes
-from sky.backends import backend_utils
 from sky.provision.kubernetes import network_utils
 from sky.utils import common_utils
 from sky.utils import env_options
@@ -786,9 +784,9 @@ def get_ssh_proxy_command(
         vars_to_fill = {
             'ssh_jump_name': ssh_jump_name,
         }
-        backend_utils.fill_template(port_fwd_proxy_cmd_template,
-                                    vars_to_fill,
-                                    output_path=port_fwd_proxy_cmd_path)
+        common_utils.fill_template(port_fwd_proxy_cmd_template,
+                                   vars_to_fill,
+                                   output_path=port_fwd_proxy_cmd_path)
         ssh_jump_proxy_command = construct_ssh_jump_command(
             private_key_path,
             ssh_jump_ip,
@@ -999,6 +997,7 @@ def clean_zombie_ssh_jump_pod(namespace: str, node_id: str):
 
 def fill_ssh_jump_template(ssh_key_secret: str, ssh_jump_image: str,
                            ssh_jump_name: str, service_type: str) -> Dict:
+    import sky
     template_path = os.path.join(sky.__root_dir__, 'templates',
                                  'kubernetes-ssh-jump.yml.j2')
     if not os.path.exists(template_path):
